// src/entities/environment/sprites/pistons/piston-types.ts
export type ThinPistonLayer =
  | "piston_bounds"
  | "head"
  | "shaft"
  | "body"
  | "impact_effects"
  | "contrails";
export type ThinPistonAnimation =
  | "loading"
  | "extending"
  | "impacting"
  | "retracting";

export type WidePistonLayer =
  | "piston_bounds"
  | "head"
  | "shaft"
  | "body"
  | "impact_effects"
  | "contrails";
export type WidePistonAnimation =
  | "loading"
  | "extending"
  | "impacting"
  | "retracting";

export type SpikePistonLayer =
  | "piston_bounds"
  | "head"
  | "shaft"
  | "body"
  | "spikes"
  | "impact_effects"
  | "contrails";
export type SpikePistonAnimation = "loading" | "extending" | "impacting";